requirements.txt
fastapi==0.103.2
numpy==1.26.1
python-dotenv==1.0.0
Requests==2.31.0
sentence_transformers==2.2.2
uvicorn==0.23.2
